4.4 faststart changes; merge missing 4.3 branch updates to master

+++ recipes/eucanetd.rb
@@ -4,11 +4,24 @@ require 'chef/version_constraint'
 include_recipe "eucalyptus::default"
 
 # Remove default virsh network which runs its own dhcp server
-execute 'virsh net-destroy default' do
-  ignore_failure true
+if Chef::VersionConstraint.new("~> 6.0").include?(node['platform_version'])
+  execute 'virsh net-destroy default' do
+    ignore_failure true
+  end
+  execute 'virsh net-autostart default --disable' do
+    ignore_failure true
+  end
 end
-execute 'virsh net-autostart default --disable' do
-  ignore_failure true
+if Chef::VersionConstraint.new("~> 7.0").include?(node['platform_version'])
+  execute 'virsh net-destroy default' do
+    ignore_failure true
+  end
+  execute 'virsh net-autostart default --disable' do
+    ignore_failure true
+  end
+  execute 'virsh net-undefine default' do
+    ignore_failure true
+  end
 end
 
 if node["eucalyptus"]["install-type"] == "packages"
@@ -37,9 +50,11 @@ if Chef::VersionConstraint.new("~> 6.0").include?(node['platform_version'])
   end
 end
 if Chef::VersionConstraint.new("~> 7.0").include?(node['platform_version'])
-  execute "Configure kernel parameters from 70-eucanetd.conf" do
-    command "/usr/lib/systemd/systemd-sysctl 70-eucanetd.conf"
-    notifies :run, "execute[Run systemd-modules-load]", :before
+  if node["eucalyptus"]["network"]["mode"] != "VPCMIDO"
+    execute "Configure kernel parameters from 70-eucanetd.conf" do
+      command "/usr/lib/systemd/systemd-sysctl 70-eucanetd.conf"
+      notifies :run, "execute[Ensure bridge modules loaded into the kernel on NC]", :before
+    end
   end
 end
